split the create employee form to a new component
