+++ components/UserCard.tsx
@@ -1,14 +1,32 @@
 import React from "react";
+
+import {
+  Select,
+  SelectItem,
+  SelectValue,
+  SelectTrigger,
+  SelectContent,
+} from "@/components/ui/select";
+
 import { User } from "../app/types/user";
 
 interface UserCardProps {
   user: User;
   isAdmin: boolean;
   isLoading: boolean;
+  selectedRole: string;
+  onRoleChange: (role: string) => void;
   onRecruit: () => void;
 }
 
-const UserCard = ({ user, isAdmin, isLoading, onRecruit }: UserCardProps) => {
+const UserCard = ({
+  user,
+  isAdmin,
+  isLoading,
+  selectedRole,
+  onRoleChange,
+  onRecruit,
+}: UserCardProps) => {
   return (
     <div className="bg-white p-6 rounded-2xl shadow-md">
       <h2 className="text-xl font-semibold">
@@ -22,7 +40,21 @@ const UserCard = ({ user, isAdmin, isLoading, onRecruit }: UserCardProps) => {
       </div>
 
       {isAdmin && (
-        <div className="flex gap-4">
+        <div className="space-y-3">
+          <Select
+            value={selectedRole}
+            onValueChange={onRoleChange}
+            disabled={isLoading}
+          >
+            <SelectTrigger className="w-full">
+              <SelectValue placeholder="Select role" />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectItem value="Doctor">Doctor</SelectItem>
+              <SelectItem value="Staff">Staff</SelectItem>
+            </SelectContent>
+          </Select>
+
           <button
             className={`w-full px-4 py-2 rounded-lg ${
               user.role === "Doctor"
